# template/stack/apps/server/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Notes API — the demo vertical slice. Regenerate openapi.json via `pnpm --filter server openapi:emit`.",
    "title": "server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/notes/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Note deleted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "bad_request",
                            "unauthorized",
                            "not_found",
                            "payload_too_large",
                            "version_skew",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "maxLength": 1024,
                          "minLength": 1,
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Request validation failed (envelope code bad_request)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "bad_request",
                            "unauthorized",
                            "not_found",
                            "payload_too_large",
                            "version_skew",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "maxLength": 1024,
                          "minLength": 1,
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Missing or unverifiable bearer token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "bad_request",
                            "unauthorized",
                            "not_found",
                            "payload_too_large",
                            "version_skew",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "maxLength": 1024,
                          "minLength": 1,
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "No such note visible to this user"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "bad_request",
                            "unauthorized",
                            "not_found",
                            "payload_too_large",
                            "version_skew",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "maxLength": 1024,
                          "minLength": 1,
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Client major version does not match the server (code version_skew)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "bad_request",
                            "unauthorized",
                            "not_found",
                            "payload_too_large",
                            "version_skew",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "maxLength": 1024,
                          "minLength": 1,
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Unexpected server error — correlate via error.requestId"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "body": {
                      "maxLength": 20000,
                      "type": "string"
                    },
                    "createdAt": {
                      "maxLength": 64,
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}[T ]\\d{2}:\\d{2}:\\d{2}",
                      "type": "string"
                    },
                    "embedding": {
                      "items": {
                        "type": "number"
                      },
                      "type": [
                        "array",
                        "null"
                      ]
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "ownerId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "sourceConfidence": {
                      "maximum": 1,
                      "minimum": 0,
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "sourceModel": {
                      "maxLength": 128,
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "title": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    }
                  },
                  "required": [
                    "body",
                    "createdAt",
                    "embedding",
                    "id",
                    "ownerId",
                    "sourceConfidence",
                    "sourceModel",
                    "title"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The requested note"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "bad_request",
                            "unauthorized",
                            "not_found",
                            "payload_too_large",
                            "version_skew",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "maxLength": 1024,
                          "minLength": 1,
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Request validation failed (envelope code bad_request)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "bad_request",
                            "unauthorized",
                            "not_found",
                            "payload_too_large",
                            "version_skew",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "maxLength": 1024,
                          "minLength": 1,
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Missing or unverifiable bearer token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "bad_request",
                            "unauthorized",
                            "not_found",
                            "payload_too_large",
                            "version_skew",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "maxLength": 1024,
                          "minLength": 1,
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "No such note visible to this user"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "bad_request",
                            "unauthorized",
                            "not_found",
                            "payload_too_large",
                            "version_skew",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "maxLength": 1024,
                          "minLength": 1,
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Client major version does not match the server (code version_skew)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "bad_request",
                            "unauthorized",
                            "not_found",
                            "payload_too_large",
                            "version_skew",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "maxLength": 1024,
                          "minLength": 1,
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Unexpected server error — correlate via error.requestId"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
...
}